Dynamically reload shaders
Pretty cool. The order of operations is hardcoded in the code. We don't use a
dependency tree because that would probably require sharing of data. The
dependencies have to be coded in some way anyway. With the right abstractions,
it can be quite clear in the code what has to happen (not quite the case right
now). It might be harder to maintain.