Fix clustered shading
We can't do light computations in CLS space easily because it has a non-uniform
scaling applied to it. Revert to having a separate attribute and allow flexibly
choosing the light space. This means we do compute the light positions twice but
at least we no longer depend on the light space and buffer directly.