WIP Darn you transition camera
So for GPU based light assignment we need to render to a depth texture. This is
still true when we are debugging and want to look at the clustering from another
view. When we aren't however, we shouldn't do the depth rendering twice. The
question is, how do we deduplicate the depth prepass?