WIP I need to sleep on this
I want to add an element to a vec while I hold a reference to an element in that
vec. This is not possible because pushing an element might re-locate all
elements when the vec needs to grow.

How can we solve this?

I was trying to use an Rc to add an indirection. However this doesn't work
because the verify function from the incremental library wants to return a
reference with a lifetime of the branch, which is tied to the Vec rather than
the Rc.

Perhaps we should move back to a two-phase computation again :').

If the computation happens in two phases we can first recompute all of our
branches with mutable access and then obtain a reference to the hopefully
now up-to-date value.

The issues with the two phases is that it is easier for the programmer to forget
to mark a dependency. Properly incorporating RefCell and Rc's everywhere isn't
easy either.

Why is this so hard???