Dang did I mention Markus is good
So we narrowed down the problem to our light indices being wrong. It turns out I
failed to realize that when we read from shared memory in a loop, we need to
make sure that the next iteration doesn't overwrite the value we are reading
before we do.

Have to apply this to my other compute shaders.